+++ src/main/java/org/br/foodjet/repository/entity/Inventory.java
@@ -2,14 +2,10 @@ package org.br.foodjet.repository.entity;
 
 import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
 import io.quarkus.runtime.annotations.RegisterForReflection;
-import java.util.List;
+import java.math.BigInteger;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToMany;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import lombok.Data;
 
@@ -17,7 +13,7 @@ import lombok.Data;
 @RegisterForReflection
 @Entity
 @Table(name = "inventory")
-public class  Inventory extends PanacheEntityBase {
+public class Inventory extends PanacheEntityBase {
 
     @Id
     @Column(name = "id")
@@ -27,5 +23,5 @@ public class  Inventory extends PanacheEntityBase {
     private String name;
 
     @Column(name = "quantity")
-    private Long quantity;
+    private BigInteger quantity;
 }
