+++ src/main/java/org/br/foodjet/resource/response/BurguerInventoryResponse.java
@@ -1,6 +1,7 @@
 package org.br.foodjet.resource.response;
 
 import io.quarkus.runtime.annotations.RegisterForReflection;
+import java.math.BigInteger;
 import lombok.Data;
 import org.br.foodjet.repository.entity.Inventory;
 
@@ -11,5 +12,5 @@ public class BurguerInventoryResponse {
     private Long id;
     private BurguerResponse burguer;
     private Inventory inventory;
-    private Long quantity;
+    private BigInteger quantity;
 }
