+++ src/main/java/org/br/foodjet/resource/response/InventoryResponse.java
@@ -1,6 +1,7 @@
 package org.br.foodjet.resource.response;
 
 import io.quarkus.runtime.annotations.RegisterForReflection;
+import java.math.BigInteger;
 import lombok.Data;
 
 @Data
@@ -9,5 +10,5 @@ public class InventoryResponse {
 
     private Long id;
     private String name;
-    private Long quantity;
+    private BigInteger quantity;
 }
